Touch temporary file after a name is created
There's an issue when trying to run virtualenvwrapper.sh on windows with
MinGw. The full issue was posted to
http://stackoverflow.com/questions/24478623/error-virtualenvwrapper-in-gitbash/28522399

Basically, it's either windows or mktemp that's not creating the tmp
file but just returning the full path of it, so just "touch"ing the file
seems to solve this issue.

Based on work by Abelardo Mendoza in:
https://bitbucket.org/dhellmann/virtualenvwrapper/pull-request/24/touch-tmp-file-to-force-windows-to-create/diff

Change-Id: I4c788686ed23c31facd7fd5538f5fabe86849a8a